Changed id types from int to long (just to be safe)

+++ iTunesSearch.Library/Models/TVEpisode.cs
@@ -8,13 +8,13 @@ namespace iTunesSearch.Library.Models
     public class TVEpisode
     {
         [DataMember(Name = "artistId")]
-        public int ShowId { get; set; }
+        public long ShowId { get; set; }
 
         [DataMember(Name = "collectionId")]
-        public int SeasonId { get; set; }
+        public long SeasonId { get; set; }
 
         [DataMember(Name = "trackId")]
-        public int EpisodeId { get; set; }
+        public long EpisodeId { get; set; }
 
         [DataMember(Name = "artistName")]
         public string ShowName { get; set; }
